- add pixar room variants - put uploadTexture function in another file
